requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.3.0
django-filter==2.2.0
django-heroku==0.3.1
django-rest-knox==4.1.0
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.13.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
